Moved new classes to a factory package, added license headers, improved some docs
