+++ src/main/java/net/dv8tion/jda/core/requests/factory/IGatewayProviderFactory.java
@@ -1,0 +1,40 @@
+/*
+ *     Copyright 2015-2017 Austin Keener & Michael Ritter & Florian Spieß
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package net.dv8tion.jda.core.requests.factory;
+
+import net.dv8tion.jda.core.JDA;
+
+/**
+ * Factory interface for the creation of new {@link IGatewayProvider IGatewayProvider} objects.
+ * <br>JDA, by default, uses {@link DefaultGatewayProviderFactory DefaultGatewayProviderFactory} for the
+ * creation of its gateway providers.
+ * <p>
+ * Implementations of this interface are provided to
+ * {@link net.dv8tion.jda.core.JDABuilder#setGatewayProviderFactory(IGatewayProviderFactory) JDABuilder.setGatewayProviderFactory(IGatewayProviderFactory)}.
+ */
+public interface IGatewayProviderFactory
+{
+    /**
+     * Called by JDA's websocket client when it's first connecting and needs an {@link IGatewayProvider IGatewayProvider}
+     * to provide it the URL it should connect to.
+     *
+     * @param jda
+     *        The JDA instance that needs an {@link IGatewayProvider IGatewayProvider}.
+     *
+     * @return A new gateway provider to provide JDA the URL it should connect to.
+     */
+    IGatewayProvider createGatewayProvider(JDA jda);
+}
